# Add nodal mask