--- a/src/DataLayouts/copyto.jl
+++ b/src/DataLayouts/copyto.jl
@@ -12,40 +12,43 @@ if VERSION ≥ v"1.11.0-beta"
     function Base.copyto!(
         dest::AbstractData{S},
         bc::Union{AbstractData, Base.Broadcast.Broadcasted},
+        mask = NoMask(),
     ) where {S}
-        Base.copyto!(dest, bc, device_dispatch(parent(dest)))
-        call_post_op_callback() && post_op_callback(dest, dest, bc)
+        Base.copyto!(dest, bc, device_dispatch(parent(dest)), mask)
+        call_post_op_callback() && post_op_callback(dest, dest, bc, mask)
         dest
     end
 else
     function Base.copyto!(
         dest::AbstractData{S},
         bc::Union{AbstractData, Base.Broadcast.Broadcasted},
+        mask = NoMask(),
     ) where {S}
         dev = device_dispatch(parent(dest))
         if dev isa ToCPU &&
            has_uniform_datalayouts(bc) &&
            dest isa EndsWithField &&
-           !(dest isa DataF)
+           !(dest isa DataF) &&
+           mask isa NoMask
             # Specialize on linear indexing when possible:
             bc′ = Base.Broadcast.instantiate(to_non_extruded_broadcasted(bc))
             @inbounds @simd for I in 1:get_N(UniversalSize(dest))
                 dest[I] = convert(S, bc′[I])
             end
         else
-            Base.copyto!(dest, bc, device_dispatch(parent(dest)))
+            Base.copyto!(dest, bc, device_dispatch(parent(dest)), mask)
         end
-        call_post_op_callback() && post_op_callback(dest, dest, bc)
+        call_post_op_callback() && post_op_callback(dest, dest, bc, mask)
         return dest
     end
 end
 
-# Specialize on non-Broadcasted objects
-function Base.copyto!(dest::D, src::D) where {D <: AbstractData}
-    copyto!(parent(dest), parent(src))
-    call_post_op_callback() && post_op_callback(dest, dest, src)
-    return dest
-end
+# This is not well optimized
+# function Base.copyto!(dest::D, src::D) where {D <: AbstractData}
+#     copyto!(parent(dest), parent(src))
+#     call_post_op_callback() && post_op_callback(dest, dest, src)
+#     return dest
+# end
 
 # broadcasting scalar assignment
 # Performance optimization for the common identity scalar case: dest .= val
@@ -53,6 +56,7 @@ function Base.copyto!(
     dest::AbstractData,
     bc::Base.Broadcast.Broadcasted{Style},
     to::AbstractDispatchToDevice,
+    mask = NoMask(),
 ) where {
     Style <:
     Union{Base.Broadcast.AbstractArrayStyle{0}, Base.Broadcast.Style{Tuple}},
@@ -61,20 +65,27 @@ function Base.copyto!(
         Base.Broadcast.Broadcasted{Style}(bc.f, bc.args, ()),
     )
     @inbounds bc0 = bc[]
-    fill!(dest, bc0)
-    call_post_op_callback() && post_op_callback(dest, dest, bc, to)
+    fill!(dest, bc0, mask)
+    call_post_op_callback() && post_op_callback(dest, dest, bc, to, mask)
 end
 
 #####
 ##### DataLayouts
 #####
 
+should_compute(::NoMask, idx) = true
+should_compute(mask::DataLayouts.AbstractMask, idx) =
+    !iszero(mask.is_active[idx])
+
 function Base.copyto!(
     dest::DataF{S},
     bc::BroadcastedUnionDataF{S, A},
     ::ToCPU,
+    mask = NoMask(),
 ) where {S, A}
-    @inbounds dest[] = convert(S, bc[])
+    if mask isa NoMask || mask[]
+        @inbounds dest[] = convert(S, bc[])
+    end
     return dest
 end
 
@@ -82,10 +93,12 @@ function Base.copyto!(
     dest::Union{IJFH{S, Nij}, IJHF{S, Nij}},
     bc::Union{BroadcastedUnionIJFH{S, Nij}, BroadcastedUnionIJHF{S, Nij}},
     ::ToCPU,
+    mask = NoMask(),
 ) where {S, Nij}
     (_, _, _, _, Nh) = size(dest)
     @inbounds for h in 1:Nh, j in 1:Nij, i in 1:Nij
         idx = CartesianIndex(i, j, 1, 1, h)
+        should_compute(mask, idx) || continue
         dest[idx] = convert(S, bc[idx])
     end
     return dest
@@ -95,10 +108,12 @@ function Base.copyto!(
     dest::Union{IFH{S, Ni}, IHF{S, Ni}},
     bc::Union{BroadcastedUnionIFH{S, Ni}, BroadcastedUnionIHF{S, Ni}},
     ::ToCPU,
+    mask = NoMask(),
 ) where {S, Ni}
     (_, _, _, _, Nh) = size(dest)
     @inbounds for h in 1:Nh, i in 1:Ni
         idx = CartesianIndex(i, 1, 1, 1, h)
+        should_compute(mask, idx) || continue
         dest[idx] = convert(S, bc[idx])
     end
     return dest
@@ -109,9 +124,11 @@ function Base.copyto!(
     dest::IJF{S, Nij},
     bc::BroadcastedUnionIJF{S, Nij, A},
     ::ToCPU,
+    mask = NoMask(),
 ) where {S, Nij, A}
     @inbounds for j in 1:Nij, i in 1:Nij
         idx = CartesianIndex(i, j, 1, 1, 1)
+        should_compute(mask, idx) || continue
         dest[idx] = convert(S, bc[idx])
     end
     return dest
@@ -121,9 +138,11 @@ function Base.copyto!(
     dest::IF{S, Ni},
     bc::BroadcastedUnionIF{S, Ni, A},
     ::ToCPU,
+    mask = NoMask(),
 ) where {S, Ni, A}
     @inbounds for i in 1:Ni
         idx = CartesianIndex(i, 1, 1, 1, 1)
+        should_compute(mask, idx) || continue
         dest[idx] = convert(S, bc[idx])
     end
     return dest
@@ -134,9 +153,11 @@ function Base.copyto!(
     dest::IF{S, Ni},
     bc::Base.Broadcast.Broadcasted{IFStyle{Ni, A}},
     ::ToCPU,
+    mask = NoMask(),
 ) where {S, Ni, A}
     @inbounds for i in 1:Ni
         idx = CartesianIndex(i, 1, 1, 1, 1)
+        should_compute(mask, idx) || continue
         dest[idx] = convert(S, bc[idx])
     end
     return dest
@@ -147,9 +168,11 @@ function Base.copyto!(
     dest::VF{S, Nv},
     bc::BroadcastedUnionVF{S, Nv, A},
     ::ToCPU,
+    mask = NoMask(),
 ) where {S, Nv, A}
     @inbounds for v in 1:Nv
         idx = CartesianIndex(1, 1, 1, v, 1)
+        should_compute(mask, idx) || continue
         dest[idx] = convert(S, bc[idx])
     end
     return dest
@@ -159,11 +182,13 @@ function Base.copyto!(
     dest::Union{VIFH{S, Nv, Ni}, VIHF{S, Nv, Ni}},
     bc::Union{BroadcastedUnionVIFH{S, Nv, Ni}, BroadcastedUnionVIHF{S, Nv, Ni}},
     ::ToCPU,
+    mask = NoMask(),
 ) where {S, Nv, Ni}
     # copy contiguous columns
     (_, _, _, _, Nh) = size(dest)
     @inbounds for h in 1:Nh, i in 1:Ni, v in 1:Nv
         idx = CartesianIndex(i, 1, 1, v, h)
+        should_compute(mask, idx) || continue
         dest[idx] = convert(S, bc[idx])
     end
     return dest
@@ -176,11 +201,13 @@ function Base.copyto!(
         BroadcastedUnionVIJHF{S, Nv, Nij},
     },
     ::ToCPU,
+    mask = NoMask(),
 ) where {S, Nv, Nij}
     # copy contiguous columns
     (_, _, _, _, Nh) = size(dest)
     @inbounds for h in 1:Nh, j in 1:Nij, i in 1:Nij, v in 1:Nv
         idx = CartesianIndex(i, j, 1, v, h)
+        should_compute(mask, idx) || continue
         dest[idx] = convert(S, bc[idx])
     end
     return dest