--- a/src/Spaces/spectralelement.jl
+++ b/src/Spaces/spectralelement.jl
@@ -116,7 +116,6 @@ function issubspace(
     return grid(hspace) === grid(level_space).full_grid.horizontal_grid
 end
 
-
 """
     SpectralElementSpaceSlab <: AbstractSpace