This PR adds a nodal mask into the `SpectralElementGrid2D`, and allows users to update it with `set_mask!(::Function, space::AbstractSpace)`. This is an initial step towards this [SDI](https://github.com/CliMA/ClimaLand.jl/issues/488).

I imagine that just the pointwise and stencil operations will go a long way, but we may need to iterate to find out what else is needed.

Note to self: I think we also need to thread the mask through in some `materialize!` calls for implicit pieces, assuming that they call `copyto!`

Supersedes #2085

Here is a rough summary of how this works

### Summary

 - We add the mask, in the form of a DataLayout to the horizontal grid, or a `DataLayouts.NoMask` object
   - Using different types like this allows us to dispatch on behavior when the mask is present and absent, effectively becoming a no-op when the mask is absent
 - We then extract and pass that mask to our kernels (i.e., `Base.copyto!`)
 - Inside `Base.copyto!`, if the mask is a datalayout, we check if `mask == 0` for a particular `i, j, f, v, h`, before doing the broadcast assignment (`dest[I] = bc[I]`) then we skip, via `continue` (or `return nothing`, in the case of cuda kernels).