@charleskawczynski 2025-03-03
I'm on the fence regarding the mask index maps:

```julia
main
julia> using Revise; include("test/standalone/Soil/mask_test.jl")
BenchmarkTools.Trial: 1559 samples with 1 evaluation per sample.
 Range (min … max):  1.176 ms …   6.669 ms  ┊ GC (min … max): 0.00% … 0.00%
 Time  (median):     3.243 ms               ┊ GC (median):    0.00%
 Time  (mean ± σ):   3.206 ms ± 343.094 μs  ┊ GC (mean ± σ):  0.00% ± 0.00%

                                                      ▅█▅▆▅▄▂  
  ▄▄▄▆▇▅▆▁▄▅▄▄▆▁▄▄▄▅▄▁▆▅▄▁▄▄▁▁▁▁▁▁▁▄▆▄▄▅▅█▅▇▆▁▁▄▁▁▁▄▆████████ █
  1.18 ms      Histogram: log(frequency) by time      3.46 ms <

 Memory estimate: 6.14 KiB, allocs estimate: 243.

with mask
julia> using Revise; include("test/standalone/Soil/mask_test.jl")
BenchmarkTools.Trial: 1606 samples with 1 evaluation per sample.
 Range (min … max):  1.006 ms …   3.441 ms  ┊ GC (min … max): 0.00% … 0.00%
 Time  (median):     3.154 ms               ┊ GC (median):    0.00%
 Time  (mean ± σ):   3.112 ms ± 332.086 μs  ┊ GC (mean ± σ):  0.00% ± 0.00%

                                                      ▅█▆▅▆▅▃  
  ▄▄▁▁▅▆▇▆▄▄▆▁▄▄▁▁▄▆▄▄▄▄▄▄▄▅▁▄▁▁▄▁▁▄▁▆▆▄▅▄█▆▇▁▁▄▄▁▄▄▄▆███████ █
  1.01 ms      Histogram: log(frequency) by time      3.35 ms <

 Memory estimate: 10.52 KiB, allocs estimate: 523.
```

The performance characteristics with the mask index maps could be pretty unpredictable without some work. Should we just do the Boolean mask? In theory I imagine that the mask index maps could be made faster, but until more profiling and work is put into it, performance could degrade in different ways.

I'm happy to leave the arrays in there and leave them as comments / todos, if people think that's a good idea. Or open an issue and revert it.