@kmdeck 2025-03-05
- From a user point of view, I dont necessarily think the mask needs to be made in the same step as the space itself.  ClimaLand has already abstracted a lot of things from the user with respect to making our "domain" (topology...etc) , this would just be one other piece in those functions.  I also dont quite see how it is possible without making ClimaCore depend on ClimaUtilities (when the mask comes from data, and CU handles the interpolation).
- I agree with @Sbozzolo that we need to be careful that masked points arent used in computations, but for me my worry is more along the lines of someone using an unsupported function and the masked points affecting the results, rather than the mask being immutable and being reset. I think the only example is reductions (where the columns are no longer independent): tracking conserved quantities, `min` being used in implicit solver convergence criteria. This might be well documented here, but someone down the line working with ClimaLand wont know to check here, necessarily, theyll just assume masked points are not part of the picture at all
- I think it would be nice to show an example of how to set the mask from data (I think this is the more common use case, since we dont have analytic functions of coordinates for continents). Even in pseudo code or using a low dimensional space and a matrix of 1s and zeros.


I saw that the performance was mixed when using the mask vs not - looking into that further would be good! We saw a 2x speedup when we reduced the number of points in the simulation by half, so I think there is still a potential 3x speedup from this. Getting to that point first, and then addressing issues about the design would be my preference.