@charleskawczynski 2025-03-05
I'm going to effectively revert to https://github.com/CliMA/ClimaCore.jl/pull/2201/commits/4e1a097ed88f809b214d48edd63fdeee95c54ea0 (I'll keep doc improvements and apply suggestions from @kmdeck).