@charleskawczynski 2025-03-05
@Sbozzolo and I chatted and hashed through a few things. I've updated the docs and made a few things more clear. I've reverted to https://github.com/CliMA/ClimaCore.jl/commit/4e1a097ed88f809b214d48edd63fdeee95c54ea0 and fixed some terminology with the documentation.