@charleskawczynski 2025-03-05
Generic IO for datalayouts is not supported, and I don't want to hack in another implementation just for masks (i.e., duplicate what we have in `read_field`), so I'll followup with restarts for masks in a followup PR. (cc @Sbozzolo)

Also, that work will collide with https://github.com/CliMA/ClimaCore.jl/pull/2162, so it makes sense to split this up. I'll open an issue to track.